feat(HUD): :sparkles: Rework the to use imgui insted of SKSEMenuFramework for more control
Refactor the code in HUD for better maintence
