+++ cmake/headerlist.cmake
@@ -1,0 +1,28 @@
+set(headers
+    src/Hooks.h
+    src/UI/HudManager.h
+    src/UI/HudState.h
+    src/UI/SlotDrawer.h
+    src/UI/PopupDrawer.h
+    src/UI/Strings.h
+    src/UI/MENU.h
+    src/UI/TextureManager.h
+    src/UI/StyleConfig.h
+    src/UI/SlotAnimator.h
+    src/UI/SlotLayout.h
+    src/Config/ConfigPath.h
+    src/Config/Config.h
+    src/Config/Slots.h
+    src/Config/EquipSlots.h
+    src/Config/SpellType.h
+    src/Persistence/SpellSettingsDB.h
+    src/Persistence/SaveSpellDB.h
+    src/Input/Assign.h
+    src/Input/Input.h
+    src/State/State.h
+    src/State/Action.h
+    src/State/AnimListener.h
+    src/State/CastGuardEvents.h
+    src/State/InventoryUtil.h
+    src/State/SyntheticInput.h
+)
